Remove unused elements and move newly added to the existing class
