+++ osu.Game/Overlays/OverlayHeaderTabControl.cs
@@ -46,7 +46,7 @@ namespace osu.Game.Overlays
             public OverlayHeaderTabItem(T value)
                 : base(value)
             {
-                Text.Text = value.ToString();
+                Text.Text = value.ToString().ToLowerInvariant();
                 Text.Font = OsuFont.GetFont(size: 14);
                 Bar.ExpandedSize = 5;
             }
