+++ osu.Game/Overlays/OverlayRulesetSelector.cs
@@ -38,6 +38,8 @@ namespace osu.Game.Overlays
         {
             if (accentColour == default)
                 AccentColour = colours.Blue;
+
+            SelectTab(TabContainer.FirstOrDefault());
         }
 
         protected override TabItem<RulesetInfo> CreateTabItem(RulesetInfo value) => new OverlayRulesetTabItem(value)
