+++ osu.Game/Overlays/Rankings/RankingsHeader.cs
@@ -1,129 +1,176 @@
 ﻿// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
 // See the LICENCE file in the repository root for full licence text.
 
-using osu.Framework.Graphics.Containers;
-using osu.Framework.Graphics;
+using System.Collections.Generic;
 using osu.Framework.Allocation;
 using osu.Framework.Bindables;
+using osu.Framework.Graphics;
+using osu.Framework.Graphics.Containers;
+using osu.Framework.Graphics.Shapes;
+using osu.Game.Graphics;
+using osu.Game.Graphics.UserInterface;
 using osu.Game.Rulesets;
 using osu.Game.Users;
-using osu.Framework.Graphics.Sprites;
-using osu.Framework.Graphics.Textures;
-using osuTK;
-using osu.Game.Graphics.UserInterface;
-using System.Collections.Generic;
 
 namespace osu.Game.Overlays.Rankings
 {
-    public class RankingsHeader : CompositeDrawable
+    public class RankingsHeader : TabControlOverlayHeader<RankingsScope>
     {
-        private const int content_height = 250;
-
-        public IEnumerable<Spotlight> Spotlights
-        {
-            get => dropdown.Items;
-            set => dropdown.Items = value;
-        }
-
-        public readonly Bindable<RankingsScope> Scope = new Bindable<RankingsScope>();
         public readonly Bindable<RulesetInfo> Ruleset = new Bindable<RulesetInfo>();
         public readonly Bindable<Country> Country = new Bindable<Country>();
         public readonly Bindable<Spotlight> Spotlight = new Bindable<Spotlight>();
 
-        private readonly OsuDropdown<Spotlight> dropdown;
+        public IEnumerable<Spotlight> Spotlights
+        {
+            get => spotlightsContainer.Spotlights;
+            set => spotlightsContainer.Spotlights = value;
+        }
+
+        private OverlayRulesetSelector rulesetSelector;
+        private SpotlightsContainer spotlightsContainer;
 
         public RankingsHeader()
         {
-            RelativeSizeAxes = Axes.X;
-            AutoSizeAxes = Axes.Y;
-
-            AddInternal(new FillFlowContainer
+            HeaderInfo.Add(new CountryFilter
             {
-                RelativeSizeAxes = Axes.X,
-                AutoSizeAxes = Axes.Y,
-                Children = new Drawable[]
-                {
-                    new RankingsRulesetSelector
-                    {
-                        Anchor = Anchor.TopCentre,
-                        Origin = Anchor.TopCentre,
-                        Current = Ruleset
-                    },
-                    new Container
-                    {
-                        Anchor = Anchor.TopCentre,
-                        Origin = Anchor.TopCentre,
-                        RelativeSizeAxes = Axes.X,
-                        Height = content_height,
-                        Children = new Drawable[]
-                        {
-                            new Container
-                            {
-                                RelativeSizeAxes = Axes.Both,
-                                Masking = true,
-                                Child = new HeaderBackground(),
-                            },
-                            new FillFlowContainer
-                            {
-                                AutoSizeAxes = Axes.Y,
-                                RelativeSizeAxes = Axes.X,
-                                Direction = FillDirection.Vertical,
-                                Spacing = new Vector2(0, 20),
-                                Children = new Drawable[]
-                                {
-                                    new RankingsScopeSelector
-                                    {
-                                        Margin = new MarginPadding { Top = 10 },
-                                        Current = Scope
-                                    },
-                                    new HeaderTitle
-                                    {
-                                        Anchor = Anchor.TopCentre,
-                                        Origin = Anchor.TopCentre,
-                                        Margin = new MarginPadding { Top = 10 },
-                                        Scope = { BindTarget = Scope },
-                                        Country = { BindTarget = Country },
-                                    },
-                                    dropdown = new OsuDropdown<Spotlight>
-                                    {
-                                        Anchor = Anchor.TopCentre,
-                                        Origin = Anchor.TopCentre,
-                                        RelativeSizeAxes = Axes.X,
-                                        Width = 0.8f,
-                                        Current = Spotlight,
-                                    }
-                                }
-                            },
-                        }
-                    }
-                }
+                Country = { BindTarget = Country }
             });
+        }
+
+        [BackgroundDependencyLoader]
+        private void load(OsuColour colours)
+        {
+            BackgroundHeight = 0;
+            TitleBackgroundColour = colours.GreySeafoamDark;
+            ControlBackgroundColour = colours.GreySeafoam;
+            rulesetSelector.AccentColour = TabControl.AccentColour = colours.Lime;
         }
 
         protected override void LoadComplete()
         {
-            Scope.BindValueChanged(onScopeChanged, true);
             base.LoadComplete();
+            Current.BindValueChanged(onCurrentChanged, true);
         }
 
-        private void onScopeChanged(ValueChangedEvent<RankingsScope> scope) =>
-            dropdown.FadeTo(scope.NewValue == RankingsScope.Spotlights ? 1 : 0, 200, Easing.OutQuint);
-
-        private class HeaderBackground : Sprite
+        private void onCurrentChanged(ValueChangedEvent<RankingsScope> scope)
         {
-            public HeaderBackground()
+            if (scope.NewValue == RankingsScope.Spotlights)
             {
-                Anchor = Anchor.Centre;
-                Origin = Anchor.Centre;
-                RelativeSizeAxes = Axes.Both;
-                FillMode = FillMode.Fill;
+                spotlightsContainer.Show();
+                return;
+            }
+
+            spotlightsContainer.Hide();
+        }
+
+        protected override Drawable CreateTitleContent() => rulesetSelector = new OverlayRulesetSelector
+        {
+            Current = Ruleset,
+        };
+
+        protected override ScreenTitle CreateTitle() => new RankingsHeaderTitle
+        {
+            Scope = { BindTarget = Current }
+        };
+
+        protected override Drawable CreateContent() => spotlightsContainer = new SpotlightsContainer
+        {
+            Spotlight = { BindTarget = Spotlight }
+        };
+
+        private class SpotlightsContainer : VisibilityContainer
+        {
+            private const int height = 150;
+            private const int duration = 200;
+
+            public readonly Bindable<Spotlight> Spotlight = new Bindable<Spotlight>();
+
+            public IEnumerable<Spotlight> Spotlights
+            {
+                get => dropdown.Items;
+                set => dropdown.Items = value;
+            }
+
+            private readonly OsuDropdown<Spotlight> dropdown;
+            private readonly Container content;
+            private readonly Box background;
+
+            public SpotlightsContainer()
+            {
+                RelativeSizeAxes = Axes.X;
+                Child = content = new Container
+                {
+                    RelativeSizeAxes = Axes.X,
+                    Height = height,
+                    Alpha = 0,
+                    Children = new Drawable[]
+                    {
+                        background = new Box
+                        {
+                            RelativeSizeAxes = Axes.Both
+                        },
+                        dropdown = new OsuDropdown<Spotlight>
+                        {
+                            Anchor = Anchor.TopCentre,
+                            Origin = Anchor.TopCentre,
+                            RelativeSizeAxes = Axes.X,
+                            Width = 0.8f,
+                            Current = Spotlight,
+                            Margin = new MarginPadding { Top = 5 }
+                        }
+                    }
+                };
             }
 
             [BackgroundDependencyLoader]
-            private void load(TextureStore textures)
+            private void load(OsuColour colours)
             {
-                Texture = textures.Get(@"Headers/rankings");
+                background.Colour = colours.GreySeafoamLight;
+            }
+
+            protected override void PopOut()
+            {
+                this.ResizeHeightTo(0, duration, Easing.OutQuint);
+                content.FadeOut(duration, Easing.OutQuint);
+            }
+
+            protected override void PopIn()
+            {
+                this.ResizeHeightTo(height, duration, Easing.OutQuint);
+                content.FadeIn(duration, Easing.OutQuint);
             }
         }
+
+        private class RankingsHeaderTitle : ScreenTitle
+        {
+            public readonly Bindable<RankingsScope> Scope = new Bindable<RankingsScope>();
+
+            public RankingsHeaderTitle()
+            {
+                Title = "ranking";
+            }
+
+            [BackgroundDependencyLoader]
+            private void load(OsuColour colours)
+            {
+                AccentColour = colours.Lime;
+            }
+
+            protected override void LoadComplete()
+            {
+                base.LoadComplete();
+                Scope.BindValueChanged(scope => Section = scope.NewValue.ToString().ToLowerInvariant(), true);
+            }
+
+            protected override Drawable CreateIcon() => new ScreenTitleTextureIcon(@"Icons/rankings");
+        }
+    }
+
+    public enum RankingsScope
+    {
+        Performance,
+        Spotlights,
+        Score,
+        Country
     }
 }
